Task/pull gfs data/2584 (#2662)
* Alembic migration to add GFS to prediction_models db table
* Cronjob for GFS model data pull
* Added unit tests for noaa_gfs
* Added NOAA cronjob script to deployment workflow
* Fixed download date bug and added tests for it
* Increased mem resources for NOAA cronjob
* Just download data for whatever most recent available date is. Wind dir still not working. Unit tests for get_download_dates needs to be fixed
* Updated launch.json

---------
