+++ api/app/.env.example
@@ -42,8 +42,11 @@ REDIS_AUTH_CACHE_EXPIRY=604800
 REDIS_DAILIES_BY_STATION_CODE_CACHE_EXPIRY=3600
 # cache data downloaded from environment canada.
 REDIS_CACHE_ENV_CANADA=True
+# cache data downloaded from NOAA
+REDIS_CACHE_NOAA=True
 # how long to cache data download from environment canada for.
 REDIS_ENV_CANADA_CACHE_EXPIRY=21600
+REDIS_NOAA_CACHE_EXPIRY=21600
 # c-haines tiff output is a feature that's useful for debugging - not intended to be set to true anywhere
 # other than on a developers machine.
 C_HAINES_OUTPUT_TIFF=False
