+++ api/app/jobs/model_data.md
@@ -92,7 +92,7 @@ Therefore, to predict weather for 12:00 PST, we must fetch the following model r
 - run time 1200: hours 006 and 009
 - run time 1800: hours 000 and 003
 
-NOAA has a delay of approximately 2.5 days when publishing GFS model data. (For example, on Feb 21 2023 at 18:30 UTC, the most recently available data is for 20230219, and only the 0000 and 0600 run times are available.) Due to this delay, `get_gfs_model_run_download_urls()` in `app.jobs.noaa` generates URLs for the date 3 days prior to the current date.
+NOAA has a delay of approximately 2.5 - 3 days when publishing GFS model data. (For example, on Feb 21 2023 at 18:30 UTC, the most recently available data is for 20230219, and only the 0000 and 0600 run times are available.) Due to this delay, `get_gfs_model_run_download_urls()` in `app.jobs.noaa` generates URLs for the dates 2.5 days (60 hours) and 3 days prior to the current date, and will then attempt to download data for both of these dates. Due to our use of our `prediction_model_run_urls` table, we don't duplicate data downloads, even though the same URLS will be generated multiple times when the NOAA cronjob runs.
 
 #### Relevant Weather Variables
