+++ api/app/tests/weather_models/test_env_canada_rdps.py
@@ -11,7 +11,8 @@ from sqlalchemy.orm import Session
 from geoalchemy2.shape import from_shape
 import app.utils.time as time_utils
 import app.weather_models.process_grib
-import app.weather_models.env_canada
+import app.jobs.env_canada
+import app.jobs.common_model_fetchers
 import app.db.crud.weather_models
 from app.db.models.weather_models import (PredictionModel, ProcessedModelRunUrl,
                                           PredictionModelRunTimestamp, PredictionModelGridSubset)
@@ -58,10 +59,10 @@ def mock_database(monkeypatch):
         return rdps_prediction_model_run
 
     monkeypatch.setattr(app.weather_models.process_grib, 'get_prediction_model', mock_get_prediction_model)
-    monkeypatch.setattr(app.weather_models.env_canada, 'get_prediction_model_run_timestamp_records',
+    monkeypatch.setattr(app.jobs.common_model_fetchers, 'get_prediction_model_run_timestamp_records',
                         mock_get_rdps_prediction_model_run_timestamp_records)
-    monkeypatch.setattr(app.weather_models.env_canada, 'get_processed_file_record', mock_get_processed_file_record)
-    monkeypatch.setattr(app.weather_models.env_canada, 'get_grids_for_coordinate', mock_get_grids_for_coordinate)
+    monkeypatch.setattr(app.jobs.env_canada, 'get_processed_file_record', mock_get_processed_file_record)
+    monkeypatch.setattr(app.jobs.common_model_fetchers, 'get_grids_for_coordinate', mock_get_grids_for_coordinate)
     monkeypatch.setattr(app.db.crud.weather_models, 'get_prediction_run', mock_get_prediction_run)
 
 
@@ -77,7 +78,7 @@ def mock_get_processed_file_record(monkeypatch):
         called = True
         return None
 
-    monkeypatch.setattr(app.weather_models.env_canada, 'get_processed_file_record', get_processed_file_record)
+    monkeypatch.setattr(app.jobs.env_canada, 'get_processed_file_record', get_processed_file_record)
 
 
 @pytest.fixture()
@@ -106,8 +107,8 @@ def mock_download_fail(monkeypatch):
 def test_get_rdps_download_urls():
     """ test to see if get_download_urls methods give the correct number of urls """
     # -1 because 000 hour has no APCP_SFC_0
-    total_num_of_urls = 85 * len(app.weather_models.env_canada.GRIB_LAYERS) - 1
-    assert len(list(app.weather_models.env_canada.get_regional_model_run_download_urls(
+    total_num_of_urls = 85 * len(app.jobs.env_canada.GRIB_LAYERS) - 1
+    assert len(list(app.jobs.env_canada.get_regional_model_run_download_urls(
         time_utils.get_utc_now(), 0))) == total_num_of_urls
 
 
@@ -118,4 +119,4 @@ def test_process_rdps(mock_download,
     # All files, except one, are marked as already having been downloaded, so we expect one file to
     # be processed.
     sys.argv = ["argv", "RDPS"]
-    assert app.weather_models.env_canada.process_models() == 1
+    assert app.jobs.env_canada.process_models() == 1
