+++ api/app/tests/weather_models/test_grib_processing.feature
@@ -54,3 +54,4 @@ Feature: Grib file processing
             | [-180.075, 0.15000000000000002, 0.0, 90.075, 0.0, -0.15]            | CMC_glb_latlon.15x.15_projection_wkt.txt         | [370, 330]        | [-124.57499999999999, 40.575]             |
             | [-180.075, 0.15000000000000002, 0.0, 90.075, 0.0, -0.15]            | CMC_glb_latlon.15x.15_projection_wkt.txt         | [315, 455]        | [-132.825, 21.825000000000003]            |
             | [-180.075, 0.15000000000000002, 0.0, 90.075, 0.0, -0.15]            | CMC_glb_latlon.15x.15_projection_wkt.txt         | [427, 245]        | [-116.02499999999998, 53.325]             |
+
