+++ openshift/templates/deploy.dc.yaml
@@ -313,11 +313,21 @@ objects:
                     configMapKeyRef:
                       name: ${GLOBAL_NAME}
                       key: env.redis-cache-env-canada
+                - name: REDIS_CACHE_NOAA
+                  valueFrom:
+                    configMapKeyRef:
+                      name: ${GLOBAL_NAME}
+                      key: env.redis-cache-noaa
                 - name: REDIS_ENV_CANADA_CACHE_EXPIRY
                   valueFrom:
                     configMapKeyRef:
                       name: ${GLOBAL_NAME}
                       key: env.redis-env-canada-cache-expiry
+                - name: REDIS_NOAA_CACHE_EXPIRY
+                  valueFrom:
+                    configMapKeyRef:
+                      name: ${GLOBAL_NAME}
+                      key: env.redis-noaa-cache-expiry
                 - name: TILESERVER_POSTGRES_READ_HOST
                   valueFrom:
                     configMapKeyRef:
