+++ openshift/templates/env_canada_gdps.cronjob.yaml
@@ -75,7 +75,7 @@ objects:
                       "run",
                       "python",
                       "-m",
-                      "app.weather_models.env_canada",
+                      "app.jobs.env_canada",
                       "GDPS",
                     ]
                   env:
